Fix cross-platform build
